Move TableEntry to its own file
