+++ src/boomerang/ssl/CMakeLists.txt
@@ -11,6 +11,7 @@ list(APPEND boomerang-ssl-sources
     ssl/Register
     ssl/RTLInstDict
     ssl/RTL
+    ssl/TableEntry
 
     # exp handling
     ssl/exp/Binary
